+++ src/app/locales/en/components/content-clusters-nodes.js
@@ -5,4 +5,12 @@ export default {
   editIps: 'Edit IPs',
   cancelIpsEdit: 'Cancel editing',
   setupIpsButton: 'Confirm changes',
+  discardChanges: 'Discard changes',
+  cancelEdit: 'Cancel edition',
+  editServices: 'Edit services',
+  applyChanges: 'Apply',
+  editingUnlockedToggles: 'Currently, you can only enable OneS3 services on cluster hosts. Disabling OneS3, toggling other services, and adding or removing hosts are not available in this software release.',
+  disabledToggleReasons: {
+    workerOneS3PortConflict: 'OneS3 services deployed on this cluster use port 443. You cannot enable OneS3 on hosts that have already deployed Cluster Worker, which uses the same port (443).',
+  },
 };