+++ src/app/locales/en/components/modals/enable-s3-modal.js
@@ -1,0 +1,26 @@
+export default {
+  title: 'Enable S3 data access protocol',
+  body: {
+    intro: 'You are going to enable S3 data access protocol on the following hosts:',
+    oneS3Deployment: 'This will be done through OneS3 service deployment on the selected hosts.',
+    deploying: '<strong>Deploying a OneS3 service on the selected hosts:</strong> {{step}}...',
+  },
+  buttons: {
+    deploy: 'Deploy',
+    cancel: 'Cancel',
+  },
+  startingDeployment: 'starting OneS3 deployment',
+  unknownStep: 'initializing',
+  noTaskId: 'Deployment has been started, but the task progress is not available. Refresh the page for updates.',
+  steps: {
+    // using S3DeploymentStep enum keys, because values contain `:`
+    OneS3CreateService: 'creating OneS3 service',
+    OneS3AddServiceHost: 'adding OneS3 service host',
+    OneproviderSetClusterIps: 'setting Oneprovider cluster IPs',
+    OnepanelSetMarker: 'setting the marker',
+    OneS3Configure: 'configuring OneS3',
+    OneproviderStart: 'starting OneS3',
+    OneproviderWaitForInit: 'waiting for OneS3 initialization',
+    LetsEncryptDisable: 'disabling Let\'s Encrypt service (if exists)',
+  },
+};