+++ src/tests/integration/components/cluster-host-table-test.js
@@ -68,4 +68,168 @@ describe('Integration | Component | cluster-host-table', function () {
     expect(clusterManagerToggle, 'cluster manager toggle').to.exist;
     expect(new ToggleHelper(clusterManagerToggle).isChecked()).to.be.false;
   });
+
+  // FIXME: refactor: remove redundancy in 3 disabled locks tests
+
+  it('in edit mode has locked toggles excluding disabled S3 services without clusterWorker', async function () {
+    const hosts = [
+      ClusterHostInfo.create({
+        hostname: 'one.example.com',
+        database: true,
+        clusterWorker: false,
+        clusterManager: false,
+        oneS3: false,
+      }),
+      ClusterHostInfo.create({
+        hostname: 'two.example.com',
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        oneS3: true,
+      }),
+    ];
+    this.set('hosts', hosts);
+
+    await render(hbs`<ClusterHostTable
+      @hosts={{this.hosts}}
+      @primaryClusterManager="one.example.com"
+      @mode="edit"
+    />`);
+
+    const hostTable = find('.cluster-host-table');
+    const helper = new HostTableHelper(hostTable);
+
+    const expectedDisabled = {
+      one: {
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        primaryClusterManager: true,
+        oneS3: false,
+      },
+      two: {
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        primaryClusterManager: true,
+        oneS3: true,
+      },
+    };
+
+    for (const hostPrefix of ['one', 'two']) {
+      for (const [service, state] of Object.entries(expectedDisabled[hostPrefix])) {
+        const toggle = helper.getToggle(`${hostPrefix}.example.com`, service);
+        const toggleHelper = new ToggleHelper(toggle);
+        expect(toggleHelper.isDisabled(), `${hostPrefix}, ${service}`).to.equal(state);
+      }
+    }
+  });
+
+  it('in create mode has unlocked all toggles', async function () {
+    const hosts = [
+      ClusterHostInfo.create({
+        hostname: 'one.example.com',
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        oneS3: false,
+      }),
+      ClusterHostInfo.create({
+        hostname: 'two.example.com',
+        database: false,
+        clusterWorker: false,
+        clusterManager: false,
+        oneS3: true,
+      }),
+    ];
+    this.set('hosts', hosts);
+
+    await render(hbs`<ClusterHostTable
+      @hosts={{this.hosts}}
+      @primaryClusterManager="one.example.com"
+      @mode="create"
+    />`);
+
+    const hostTable = find('.cluster-host-table');
+    const helper = new HostTableHelper(hostTable);
+
+    const expectedDisabled = {
+      one: {
+        database: false,
+        clusterWorker: false,
+        clusterManager: false,
+        primaryClusterManager: false,
+        oneS3: false,
+      },
+      two: {
+        database: false,
+        clusterWorker: false,
+        clusterManager: false,
+        primaryClusterManager: false,
+        oneS3: false,
+      },
+    };
+
+    for (const hostPrefix of ['one', 'two']) {
+      for (const [service, state] of Object.entries(expectedDisabled[hostPrefix])) {
+        const toggle = helper.getToggle(`${hostPrefix}.example.com`, service);
+        const toggleHelper = new ToggleHelper(toggle);
+        expect(toggleHelper.isDisabled(), `${hostPrefix}, ${service}`).to.equal(state);
+      }
+    }
+  });
+
+  it('in show mode has locked all toggles', async function () {
+    const hosts = [
+      ClusterHostInfo.create({
+        hostname: 'one.example.com',
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        oneS3: false,
+      }),
+      ClusterHostInfo.create({
+        hostname: 'two.example.com',
+        database: false,
+        clusterWorker: false,
+        clusterManager: false,
+        oneS3: true,
+      }),
+    ];
+    this.set('hosts', hosts);
+
+    await render(hbs`<ClusterHostTable
+      @hosts={{this.hosts}}
+      @primaryClusterManager="one.example.com"
+      @mode="show"
+    />`);
+
+    const hostTable = find('.cluster-host-table');
+    const helper = new HostTableHelper(hostTable);
+
+    const expectedDisabled = {
+      one: {
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        primaryClusterManager: true,
+        oneS3: true,
+      },
+      two: {
+        database: true,
+        clusterWorker: true,
+        clusterManager: true,
+        primaryClusterManager: true,
+        oneS3: true,
+      },
+    };
+
+    for (const hostPrefix of ['one', 'two']) {
+      for (const [service, state] of Object.entries(expectedDisabled[hostPrefix])) {
+        const toggle = helper.getToggle(`${hostPrefix}.example.com`, service);
+        const toggleHelper = new ToggleHelper(toggle);
+        expect(toggleHelper.isDisabled(), `${hostPrefix}, ${service}`).to.equal(state);
+      }
+    }
+  });
 });